Implement disk space check and file download functionality; refactor existing methods for better modularity
