+++ src/http/handler/geodat.go
@@ -13,7 +13,6 @@ import (
 	"time"
 
 	"github.com/daniellavrushin/b4/log"
-	"golang.org/x/sys/unix"
 )
 
 type GeodatDownloadRequest struct {
@@ -345,73 +344,6 @@ func (api *API) handleGeodatUpload(w http.ResponseWriter, r *http.Request) {
 	})
 }
 
-func checkDiskSpace(dir string, needed int64) error {
-	var stat unix.Statfs_t
-	if err := unix.Statfs(dir, &stat); err != nil {
-		return fmt.Errorf("failed to check disk space on %s: %v", dir, err)
-	}
-	available := int64(stat.Bavail) * int64(stat.Bsize)
-	if available < needed {
-		availMB := float64(available) / (1024 * 1024)
-		neededMB := float64(needed) / (1024 * 1024)
-		return fmt.Errorf("not enough disk space in %s: %.1f MB available, need %.1f MB", dir, availMB, neededMB)
-	}
-	return nil
-}
-
-func downloadFile(url, destPath string) (int64, error) {
-	resp, err := http.Get(url)
-	if err != nil {
-		return 0, fmt.Errorf("failed to fetch %s: %v", url, err)
-	}
-	defer resp.Body.Close()
-
-	if resp.StatusCode != http.StatusOK {
-		return 0, fmt.Errorf("remote server returned %s for %s", resp.Status, url)
-	}
-
-	dir := filepath.Dir(destPath)
-
-	if resp.ContentLength > 0 {
-		if err := checkDiskSpace(dir, resp.ContentLength); err != nil {
-			return 0, err
-		}
-	}
-
-	tmpFile, err := os.CreateTemp(dir, ".geodat-download-*.tmp")
-	if err != nil {
-		return 0, fmt.Errorf("failed to create temp file in %s: %v", dir, err)
-	}
-	tmpPath := tmpFile.Name()
-
-	cleanup := func() {
-		tmpFile.Close()
-		os.Remove(tmpPath)
-	}
-
-	size, err := io.Copy(tmpFile, resp.Body)
-	if err != nil {
-		cleanup()
-		return 0, fmt.Errorf("failed to write data to disk (%d bytes written): %v", size, err)
-	}
-
-	if err := tmpFile.Sync(); err != nil {
-		cleanup()
-		return 0, fmt.Errorf("failed to flush data to disk: %v", err)
-	}
-
-	if err := tmpFile.Close(); err != nil {
-		os.Remove(tmpPath)
-		return 0, fmt.Errorf("failed to finalize file write: %v", err)
-	}
-
-	if err := os.Rename(tmpPath, destPath); err != nil {
-		os.Remove(tmpPath)
-		return 0, fmt.Errorf("failed to move downloaded file to %s: %v", destPath, err)
-	}
-
-	return size, nil
-}
 
 // @Summary Get geodat file info
 // @Tags Geodat
