+++ src/http/handler/system.go
@@ -268,10 +268,18 @@ func (api *API) handleUpdate(w http.ResponseWriter, r *http.Request) {
 
 		extraPaths := "/opt/sbin:/opt/bin:/usr/local/sbin:/usr/local/bin:/usr/sbin:/usr/bin:/sbin:/bin"
 		currentPath := os.Getenv("PATH")
+		existing := make(map[string]struct{})
+		for _, entry := range strings.Split(currentPath, ":") {
+			existing[entry] = struct{}{}
+		}
 		fullPath := currentPath
 		for _, p := range strings.Split(extraPaths, ":") {
-			if !strings.Contains(currentPath, p) {
-				fullPath += ":" + p
+			if _, ok := existing[p]; !ok && p != "" {
+				if fullPath != "" {
+					fullPath += ":"
+				}
+				fullPath += p
+				existing[p] = struct{}{}
 			}
 		}
